pending refactoring of valueSpec processing
